backend/py/model/requirements.txt
numpy
pandas
psycopg2_binary
scikit-learn
